requirements.txt
llama-index
llama-index-readers-file
azure-common==1.1.28
azure-core==1.30.1
azure-cosmos==4.7.0
azure-functions==1.19.0
azure-identity==1.16.0
azure-mgmt-core==1.4.0
azure-mgmt-cosmosdb==9.5.0
azure-search==1.0.0b2
azure-search-documents==11.4.0
azure-storage-blob==12.20.0
azure-storage-blob-changefeed==12.0.0b5
fastapi==0.111.1
itemadapter==0.9.0
openai==1.37.1
# psycopg2==2.9.9
# pymongo==4.7.2
python-dotenv==1.0.1
# schedule==1.2.2
Scrapy==2.11.2
streamlit==1.37.0
uvicorn==0.30.3
azure-ai-documentintelligence --pre

llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index-packs-agents-coa
llama_index-packs-query-understanding-agent
pydantic-settings

azure-servicebus